+++ packages/ui/src/components/Visualization/Canvas/CanvasForm.test.tsx
@@ -9,7 +9,7 @@ import { IVisualizationNode, VisualComponentSchema } from '../../../models/visua
 import { EntitiesContext } from '../../../providers/entities.provider';
 import { SchemaService } from '../../Form';
 import { CustomAutoFieldDetector } from '../../Form/CustomAutoField';
-import { CanvasForm, getNonDefaultProperties, getNonEmptyProperties } from './CanvasForm';
+import { CanvasForm } from './CanvasForm';
 import { CanvasNode } from './canvas.models';
 
 describe('CanvasForm', () => {
@@ -189,135 +189,5 @@ describe('CanvasForm', () => {
         throw new Error(`Error rendering ${name} pattern: ${(e as any).message}`);
       }
     });
-  });
-});
-
-describe('CanvasForm getNonDefaultProperties()', () => {
-  const schema = {
-    type: 'object',
-    properties: {
-      parameters: {
-        properties: {
-          events: {
-            type: 'string',
-            default: 'CREATE,MODIFY,DELETE',
-            title: 'Events',
-          },
-          concurrentConsumers: {
-            type: 'integer',
-            default: 1,
-            title: 'Concurrent Consumers',
-          },
-          bridgeErrorHandler: {
-            type: 'boolean',
-            default: false,
-            title: 'Bridge Error Handler',
-          },
-        },
-      },
-    },
-  } as unknown as JSONSchemaType<unknown>;
-
-  const newModel: Record<string, unknown> = {
-    id: 'from-7126',
-    description: 'test',
-    steps: [],
-    uri: 'file-watch',
-    parameters: {
-      events: 'CREATE',
-      concurrentConsumers: '1',
-      bridgeErrorHandler: false,
-    },
-  };
-
-  const newModelExpected: Record<string, unknown> = {
-    id: 'from-7126',
-    description: 'test',
-    steps: [],
-    uri: 'file-watch',
-    parameters: {
-      events: 'CREATE',
-    },
-  };
-
-  it('should return only the properties which are different from default', () => {
-    const newModelClean = getNonDefaultProperties(schema?.properties.parameters.properties, newModel);
-    expect(newModelClean).toMatchObject(newModelExpected);
-  });
-});
-
-describe('CanvasForm getNonEmptyProperties()', () => {
-  const schema = {
-    type: 'object',
-    properties: {
-      parameters: {
-        properties: {
-          events: {
-            type: 'string',
-            default: 'CREATE,MODIFY,DELETE',
-            title: 'Events',
-          },
-          concurrentConsumers: {
-            type: 'integer',
-            default: 1,
-            title: 'Concurrent Consumers',
-          },
-          bridgeErrorHandler: {
-            type: 'boolean',
-            default: false,
-            title: 'Bridge Error Handler',
-          },
-          exchangePattern: {
-            type: 'object',
-            title: 'Exchange Pattern',
-          },
-        },
-      },
-    },
-  } as unknown as JSONSchemaType<unknown>;
-
-  const newModel: Record<string, unknown> = {
-    id: 'from-7126',
-    description: 'test',
-    steps: [],
-    uri: 'file-watch',
-    parameters: {
-      events: 'CREATE',
-      concurrentConsumers: '',
-      bridgeErrorHandler: false,
-      exchangePattern: {},
-    },
-  };
-
-  const newModelIntermediate: Record<string, unknown> = {
-    id: 'from-7126',
-    description: 'test',
-    steps: [],
-    uri: 'file-watch',
-    parameters: {
-      events: 'CREATE',
-      concurrentConsumers: '',
-      exchangePattern: {},
-    },
-  };
-
-  const newModelExpected: Record<string, unknown> = {
-    id: 'from-7126',
-    description: 'test',
-    steps: [],
-    uri: 'file-watch',
-    parameters: {
-      events: 'CREATE',
-    },
-  };
-
-  it('should return only the properties which are different from default', () => {
-    const newModelClean = getNonDefaultProperties(schema?.properties.parameters.properties, newModel);
-    expect(newModelClean).toMatchObject(newModelIntermediate);
-  });
-
-  it('should return only the non-empty properties', () => {
-    const newModelClean = getNonEmptyProperties(newModel);
-    expect(newModelClean).toMatchObject(newModelExpected);
   });
 });
