+++ packages/ui/src/components/Visualization/Canvas/CanvasForm.tsx
@@ -1,6 +1,5 @@
 import { AutoField, AutoFields, AutoForm, ErrorsField } from '@kaoto-next/uniforms-patternfly';
 import { Title } from '@patternfly/react-core';
-import type { JSONSchemaType } from 'ajv';
 import { FunctionComponent, useCallback, useContext, useEffect, useMemo, useRef } from 'react';
 import { EntitiesContext } from '../../../providers/entities.provider';
 import { ErrorBoundary } from '../../ErrorBoundary';
@@ -14,46 +13,6 @@ import './CanvasForm.scss';
 
 interface CanvasFormProps {
   selectedNode: CanvasNode;
-}
-
-export function getNonDefaultProperties(
-  obj1: JSONSchemaType<unknown>,
-  obj2: Record<string, unknown>,
-): Record<string, unknown> {
-  const newModelUpdated = Object.entries(obj2.parameters as object).reduce(
-    (acc: [string, unknown][], currentValue: [string, unknown]) => {
-      if (!(obj1[currentValue[0]]['default'] == currentValue[1])) {
-        acc.push(currentValue);
-      }
-      return acc;
-    },
-    [],
-  );
-  return { ...obj2, parameters: Object.fromEntries(newModelUpdated) };
-}
-
-export function getNonEmptyProperties(obj: Record<string, unknown>): Record<string, unknown> {
-  const result = Object.entries(obj.parameters as object).reduce(
-    (acc: [string, unknown][], currentValue: [string, unknown]) => {
-      switch (typeof currentValue[1]) {
-        case 'string':
-          if (currentValue[1].trim().length !== 0) {
-            acc.push(currentValue);
-          }
-          break;
-        case 'object':
-          if (Object.keys(currentValue[1] as object).length !== 0) {
-            acc.push(currentValue);
-          }
-          break;
-        default:
-          acc.push(currentValue);
-      }
-      return acc;
-    },
-    [],
-  );
-  return { ...obj, parameters: Object.fromEntries(result) };
 }
 
 const omitFields = [
@@ -93,18 +52,7 @@ export const CanvasForm: FunctionComponent<CanvasFormProps> = (props) => {
 
   const handleOnChange = useCallback(
     (newModel: Record<string, unknown>) => {
-      if (newModel.parameters === undefined) {
-        props.selectedNode.data?.vizNode?.updateModel(newModel);
-      } else {
-        // newModelNonDefault will contain only those properties that has different value than default.
-        const newModelNonDefault = getNonDefaultProperties(
-          props.selectedNode.data?.vizNode?.getComponentSchema()?.schema?.properties.parameters.properties,
-          newModel,
-        );
-        // newModelClean will contain only non empty propertis that has different value than default.
-        const newModelClean = getNonEmptyProperties(newModelNonDefault);
-        props.selectedNode.data?.vizNode?.updateModel(newModelClean);
-      }
+      props.selectedNode.data?.vizNode?.updateModel(newModel);
       entitiesContext?.updateSourceCodeFromEntities();
     },
     [entitiesContext, props.selectedNode.data?.vizNode, props.selectedNode.id],
